+++ expenses/transactionstatus.go
@@ -55,6 +55,7 @@ func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request oper
 	res := &operations.GetSyncTransactionResponse{
 		StatusCode:  httpRes.StatusCode,
 		ContentType: contentType,
+		RawResponse: httpRes,
 	}
 	switch {
 	case httpRes.StatusCode == 200:
@@ -72,7 +73,7 @@ func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request oper
 	return res, nil
 }
 
-// GetSyncTransactions - Get Sync Transactions
+// GetSyncTransactions - Get Sync transactions
 // Get's the transactions and status for a sync
 func (s *transactionStatus) GetSyncTransactions(ctx context.Context, request operations.GetSyncTransactionsRequest) (*operations.GetSyncTransactionsResponse, error) {
 	baseURL := s.serverURL
@@ -103,6 +104,7 @@ func (s *transactionStatus) GetSyncTransactions(ctx context.Context, request ope
 	res := &operations.GetSyncTransactionsResponse{
 		StatusCode:  httpRes.StatusCode,
 		ContentType: contentType,
+		RawResponse: httpRes,
 	}
 	switch {
 	case httpRes.StatusCode == 200: