+++ expenses/RELEASES.md
@@ -7,3 +7,67 @@ Based on:
 - Speakeasy CLI 1.5.3 https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [Go v0.1.0] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.1.0 - expenses
+
+## 2023-02-28 00:13:04
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.5.4 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.1.1] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.1.1 - expenses
+
+## 2023-03-01 00:16:16
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.6.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.2.0] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.2.0 - expenses
+
+## 2023-03-02 00:14:13
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.7.1 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.3.0] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.3.0 - expenses
+
+## 2023-03-03 01:52:38
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.7.1 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.3.1] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.3.1 - expenses
+
+## 2023-03-04 00:12:19
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.8.2 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.4.0] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.4.0 - expenses
+
+## 2023-03-06 00:13:21
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.8.4 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.4.1] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.4.1 - expenses
+
+## 2023-03-07 00:14:18
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.8.5 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.4.2] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.4.2 - expenses
+
+## 2023-03-08 00:13:08
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.8.6 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.4.3] https://github.com/codatio/client-sdk-go/releases/tag/expenses/v0.4.3 - expenses